+++ src/app/array/first/impl/firstOwn.ts
@@ -1,0 +1,9 @@
+import { First } from '../util/first.interface';
+
+export class FirstOwn implements First {
+
+  first<T>(array: T[]): T {
+    return array[0];
+  }
+
+}