+++ src/app/array/first/util/first.function.ts
@@ -1,0 +1,11 @@
+import { First } from './first.interface';
+
+export class FirstFunction {
+
+  constructor(private firstImpl: First) { }
+
+  first<T>(array: T[]): T {
+    return this.firstImpl.first(array);
+  }
+
+}