+++ src/app/array/first/util/first.interface.ts
@@ -1,0 +1,3 @@
+export interface First {
+  first<T>(array: T[]): T;
+}