+++ src/app/array/last/util/last.interface.ts
@@ -1,0 +1,3 @@
+export interface Last {
+  last<T>(array: T[]): T;
+}