+++ src/app/array/orderBy/util/orderBy.interface.ts
@@ -1,0 +1,3 @@
+export interface OrderBy {
+  orderBy<T>(key: string, array: T[]): T[];
+}