+++ src/app/array/reverse/impl/reverseOwn.ts
@@ -1,0 +1,9 @@
+import { Reverse } from '../util/reverse.interface';
+
+export class ReverseOwn implements Reverse {
+
+  reverse<T>(array: T[]): T[] {
+    return array.reverse();
+  }
+
+}