+++ src/app/array/reverse/util/reverse.interface.ts
@@ -1,0 +1,3 @@
+export interface Reverse {
+  reverse<T>(array: T[]): T[];
+}