+++ src/app/array/util/array.util.interface.ts
@@ -1,6 +1,0 @@
-export interface ArrayUtil {
-  first<T>(array: T[]): T;
-  last<T>(array: T[]): T;
-  orderBy<T>(key: string, array: T[]): T[];
-  reverse<T>(array: T[]): T[];
-}