+++ src/test/integration/array/first/own/index.spec.ts
@@ -1,0 +1,6 @@
+import { FirstOwn } from '../../../../../app/array/first/impl/firstOwn';
+import { FirstTest } from '../test';
+
+const firstOwn = new FirstOwn();
+
+new FirstTest(firstOwn).runTest();