+++ src/test/integration/array/first/test.ts
@@ -1,5 +1,5 @@
-import { ArrayUtil } from '../../../app/array/util/array.util.interface';
-import { ArrayFunctions } from '../../../app/array/util/array.function';
+import { First } from '../../../../app/array/first/util/first.interface';
+import { FirstFunction } from '../../../../app/array/first/util/first.function';
 
 export class FirstTest {
 
@@ -17,13 +17,13 @@ export class FirstTest {
     }
   ];
 
-  constructor(private arrayUtilImpl: ArrayUtil) {}
+  constructor(private firstImpl: First) {}
 
   runTest(): void {
     describe('Get first item from array', () => {
       this.data.forEach((item) => {
         test('should return first item from array', () => {
-          const result = new ArrayFunctions(this.arrayUtilImpl).first(item.array);
+          const result = new FirstFunction(this.firstImpl).first(item.array);
 
           expect(result).toEqual(item.expect);
         });