+++ src/test/integration/array/last/own/index.spec.ts
@@ -1,0 +1,6 @@
+import { LastOwn } from '../../../../../app/array/last/impl/lastOwn';
+import { LastTest } from '../test';
+
+const lastOwn = new LastOwn();
+
+new LastTest(lastOwn).runTest();