+++ src/test/integration/array/reverse/own/index.spec.ts
@@ -1,0 +1,6 @@
+import { ReverseOwn } from '../../../../../app/array/reverse/impl/reverseOwn';
+import { ReverseTest } from '../test';
+
+const reverseOwn = new ReverseOwn();
+
+new ReverseTest(reverseOwn).runTest();