+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+- [Array] implementation [lodas](https://github.com/lodash/lodash) to own implementation
+- [Array] divide functions to import only function on use
+
 ### Deprecated
 ### Removed
 ### Fixed